Force the use of python3
This fixes building the installer in macOS as the default python binary is python2 (included by macOS).
Moreover python2 is deprecated in macOS and the path forward is to specifically use python3 instead of the default python.
As far as I can see, pip (required to install certifi) is only available for python3 in homebrew.

Signed-off-by: Kate <kit.ty.kate@disroot.org>