Expand linux root when installing from free space
When installing from free space the available size of the free chunk was
not being passed to the OSInstaller in order to extend the Linux rootfs
partition.

This passes (free_size - stub_size) to the OSInstaller so that it can
expand the linux root to cover all available space in that chunk.

Signed-off-by: Javier Alvarez <javier.alvarez@allthingsembedded.net>