stub: Collect non-iBoot FUD firmwares into all_firmware.tar.gz
Turns out we need these for Touch Bar, and for MTP on M2 machines...

This also ends up collecting duplicate ANE>0 firmwares for some machines.
That's probably a mistake in how Apple marked them in the config as non-iBoot,
or just a sign that those ANEs are always disabled anyway, but let's not risk it
by trying to filter them out...

These are img4ps and need img4tool to extract. No proper parsing done yet, this
is just the fallback path.

Signed-off-by: Hector Martin <marcan@marcan.st>