Check if a partition is a container
The Asahi Linux installer assumes a partition whose type is APFS is a
container, but it sometimes turns otherwise because it is not formatted,
for example.

Commit b550d86cf81d64449eff776d6b201245ff3ca9eb tried to solve the issue
by filling Partition.container with a dictionary containing empty
Volumes property. However, it broke some existing code checking if
Partition.container is None. Particularly, it broke
InstallerMain.can_resize which has such check, and accesses CapacityFree
property.

It is difficult to fill Partition.container with a dummy dictionary
containing lots of properties which an actual container can have, and
it cannot prevent from perform some operation on the non-existing
container.

This change removes the dummy dictionary and instead adds code to check
if Partition.container is None to every place where assumption that a
partition is a container is made. Fortunately, there are only three such
places; namely, OSEnum.collect_recovery, OSEnum.collect_part, and
OSEnum.main_loop. OSEnum.collect_part already has such a check.

Signed-off-by: Akihiko Odaki <akihiko.odaki@gmail.com>