Initial support for USB/external installation
This seems to mostly work as far as Apple's tooling is concerned
(with some bugs), but of course m1n1 can't chainload from USB,
so it is only useful as a demo so far. Hence, it is only available
in expert mode and will only let you install the m1n1/tethered
boot option.

Signed-off-by: Hector Martin <marcan@marcan.st>