When using "external_boot" mode list all the possible OSes to install
If a user would like to install Asahi to external media and selects this
under expert mode, they are left with no list of OSes to choose from, if
they are hoping to use images from online sources to use. With these
lines removed the full list exists, just like any other install.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>