main: Increase overhead threshold to 16GB
Looks like a lot of newer installs end up with 13GB or so of overhead as
calculated by the installer. This is confusing users. We still have no
idea what this is about or where it comes from, but it's not the Time
Machine issue, so let's just squelch the message in these cases.

Signed-off-by: Hector Martin <marcan@marcan.st>