asahi_firmware.img4: Try lzfse SOVERSION 1 first
Upstream lzfse library has no SOVERSION set, but distributions
with library packaging guidelines wants them.

Fedora[1][2] is apparently already using SOVERSION 1 (and patching this
file using sed on package builds). Debian[3] is about to do the same.

Try first to load liblzfse.so.1 and if not found, fall back on the
unversioned soname as before. This should make everyone happy without
having to patch it everywhere.

[1]: https://src.fedoraproject.org/rpms/lzfse/blob/rawhide/f/60.patch
[2]: https://src.fedoraproject.org/rpms/asahi-installer/blob/fb68ade87292d3210fbe8845da829add6b1187c4/f/asahi-installer.spec#_54
[3]: https://salsa.debian.org/bananas-team/lzfse/-/blob/debian/unstable/debian/patches/0001-debian-set-library-SONAME.patch

Signed-off-by: Andreas Henriksson <andreas@fatal.se>