Avoid dividing by zero when showing progress
PackageInstaller's fdcopy does not check if its `size` argument is zero
before dividing `copied` by `size`. As a result, if non-image partitions
in the installer package (typically, the ESP) contain empty files, copying
those to disk raises a ZeroDivisionError that abruptly terminates the
installation process, leaving it in an inconsistent state.

Check if `size` is different than zero, and if it is skip showing progress
(which is totally fine and desired).

Signed-off-by: NoisyCoil <noisycoil@tutanota.com>